Move order by integration tests to dedicated class.
